+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/ActivityFactory.kt
@@ -64,9 +64,7 @@ class ActivityFactory(
         }
 
         timestamps {
-            if (modeSettings.timestampEnabled) {
-                start = modeSettings.timestampTarget.getStart(context).toEpochMilli()
-            }
+            start = modeSettings.timestampTarget.getStart(context).toEpochMilli()
         }
     }
 }