+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Icons now have rounded corners
 - Icons are now 256x256 instead of 512x512
 
+### Removed
+
+- Toggle for "show elapsed time" because Discord now automatically generates a timestamp if it's missing
+
 ## [1.9.1] - 2025-10-27
 
 ### Added